Create PlayerInteract script
Move method Interact from Player to new script PlayerInteract + Modify inputSystemAction to interact with F instead of E
