+++ Assets/Scripts/Entities/PlayerScripts/PlayerInteract.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3afdcc50cc8688c43b08a14e65bc30f9
